admin: pin AccountBlock to admin sidebar so admins can switch / sign out without leaving /admin
- Extract AccountBlock + AddAccountForm from ExplorerRail into a shared
  components/AccountBlock so both the dashboard rail and the admin
  sidebar render the exact same dropdown.
- Mount it at the bottom of the admin <aside>; switch <aside> to
  flex-col so the nav scrolls and the account block stays pinned.
- Fetch full_name on the admin layout so the trigger shows the same
  identity surface as the dashboard rail.
- Drop the leftover Invitations nav link + unused Mail/ScrollText
  imports — the Invitations panel was already wiped in the prior
  commit; this finishes the cleanup so the nav matches the wave.

Resolves: 'make sure i can log out and switch accounts from the admin
panel. Right now im not sure you can.'
